kiln/requirements.txt
asgiref==3.4.1
dbus-python==1.2.18
Django==4.0.1
djangorestframework==3.13.1
pycairo==1.20.1
PyGObject==3.42.0
pytz==2021.3
sqlparse==0.4.2
